ajusta o modulo user

+++ src/modules/user/user.resolver.ts
@@ -3,33 +3,51 @@ import { UserService } from './user.service';
 import { User } from './entities/user.entity';
 import { CreateUserInput } from './dto/create-user.input';
 import { UpdateUserInput } from './dto/update-user.input';
+import { UserInputError } from 'apollo-server-express';
 
 @Resolver(() => User)
 export class UserResolver {
   constructor(private readonly userService: UserService) {}
 
+  /** Criar usuário */
   @Mutation(() => User)
-  createUser(@Args('data') data: CreateUserInput) {
+  async createUser(@Args('data') data: CreateUserInput) {
+    if (!data.name?.trim()) {
+      throw new UserInputError('User name cannot be empty');
+    }
     return this.userService.create(data);
   }
 
+  /** Buscar todos os usuários */
   @Query(() => [User], { name: 'users' })
   findAll() {
     return this.userService.findAll();
   }
 
+  /** Buscar um usuário pelo ID */
   @Query(() => User, { name: 'user' })
-  findOne(@Args('id') id: string) {
+  async findOne(@Args('id') id: string) {
+    if (!id.match(/^[a-f0-9-]+$/i)) {
+      throw new UserInputError('Invalid user ID format');
+    }
     return this.userService.findOne(id);
   }
 
+  /** Atualizar usuário */
   @Mutation(() => User)
-  updateUser(@Args('id') id: string, @Args('data') data: UpdateUserInput) {
+  async updateUser(@Args('id') id: string, @Args('data') data: UpdateUserInput) {
+    if (data.name && !data.name.trim()) {
+      throw new UserInputError('User name cannot be empty');
+    }
     return this.userService.update(id, data);
   }
 
+  /** Remover usuário */
   @Mutation(() => User)
-  removeUser(@Args('id') id: string) {
+  async removeUser(@Args('id') id: string) {
+    if (!id.match(/^[a-f0-9-]+$/i)) {
+      throw new UserInputError('Invalid user ID format');
+    }
     return this.userService.remove(id);
   }
 }
